requirements_py311_gcp.txt
# GCP Dependencies
google-cloud-storage>=2.10.0
google-cloud-run>=0.10.0
google-cloud-core>=2.3.0

# ML Libraries (Python 3.11 compatible)
autogluon.tabular
xgboost
lightgbm
catboost
interpret

# Core Dependencies
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.64.0
psutil>=5.9.0
PyYAML>=6.0.0 